@@ -1,18 +1,11 @@
-using System.Net;
-using RestSharp.Tests.Integrated.Server;
 // ReSharper disable ClassNeverInstantiated.Local
 
 namespace RestSharp.Tests.Integrated;
 
-[Collection(nameof(TestServerCollection))]
-public class RequestFailureTests {
-    readonly RestClient        _client;
-    readonly TestServerFixture _fixture;
+using Server;
 
-    public RequestFailureTests(TestServerFixture fixture) {
-        _client  = new RestClient(fixture.Server.Url);
-        _fixture = fixture;
-    }
+public sealed class RequestFailureTests(WireMockTestServer server) : IClassFixture<WireMockTestServer>, IDisposable {
+    readonly RestClient _client = new(server.Url!);
 
     [Fact]
     public async Task Handles_GET_Request_Errors() {
@@ -25,7 +18,7 @@ public class RequestFailureTests {
     [Fact]
     public async Task Handles_GET_Request_Errors_With_Response_Type() {
         var request  = new RestRequest("status?code=404");
-        var response = await _client.ExecuteAsync<Response>(request);
+        var response = await _client.ExecuteAsync<SuccessResponse>(request);
 
         response.StatusCode.Should().Be(HttpStatusCode.NotFound);
         response.Data.Should().Be(null);
@@ -33,10 +26,10 @@ public class RequestFailureTests {
 
     [Fact]
     public async Task Throws_on_unsuccessful_call() {
-        var client  = new RestClient(new RestClientOptions(_fixture.Server.Url) { ThrowOnAnyError = true });
-        var request = new RestRequest("status?code=500");
+        using var client  = new RestClient(new RestClientOptions(server.Url!) { ThrowOnAnyError = true });
+        var       request = new RestRequest("status?code=500");
 
-        var task = () => client.ExecuteAsync<Response>(request);
+        var task = () => client.ExecuteAsync<SuccessResponse>(request);
         await task.Should().ThrowExactlyAsync<HttpRequestException>();
     }
 
@@ -61,7 +54,7 @@ public class RequestFailureTests {
     public async Task GetAsync_generic_throws_on_unsuccessful_call() {
         var request = new RestRequest("status?code=500");
 
-        var task = () => _client.GetAsync<Response>(request);
+        var task = () => _client.GetAsync<SuccessResponse>(request);
         await task.Should().ThrowExactlyAsync<HttpRequestException>();
     }
 
@@ -69,12 +62,9 @@ public class RequestFailureTests {
     public async Task GetAsync_returns_null_on_404() {
         var request = new RestRequest("status?code=404");
 
-        var response = await _client.GetAsync<Response>(request);
+        var response = await _client.GetAsync<SuccessResponse>(request);
         response.Should().BeNull();
     }
 
-    class Response {
-        // ReSharper disable once UnusedMember.Local
-        public string Message { get; set; } = null!;
-    }
+    public void Dispose() => _client.Dispose();
 }
