@@ -1,25 +1,23 @@
-﻿using System.Net;
-using RestSharp.Tests.Integrated.Server;
+﻿namespace RestSharp.Tests.Integrated;
 
-namespace RestSharp.Tests.Integrated;
-
-[Collection(nameof(TestServerCollection))]
-public class AsyncTests {
-    readonly ITestOutputHelper _output;
-    readonly RestClient        _client;
-    readonly string            _host;
-
-    public AsyncTests(TestServerFixture fixture, ITestOutputHelper output) {
-        _output = output;
-        _client = new RestClient(fixture.Server.Url);
-        _host   = _client.Options.BaseUrl!.Host;
-    }
-
-    class Response {
-        public string Message { get; set; } = null!;
-    }
+public sealed class AsyncTests(WireMockTestServer server) : IClassFixture<WireMockTestServer>, IDisposable {
+    readonly RestClient _client = new(server.Url!);
 
     [Fact]
+    public async Task Can_Handle_Exception_Thrown_By_Interceptor_BeforeDeserialization() {
+        const string exceptionMessage = "Thrown from OnBeforeDeserialization";
+
+        var request = new RestRequest("success") {
+            Interceptors = [new ThrowingInterceptor(exceptionMessage)]
+        };
+
+        var response = await _client.ExecuteAsync<Response>(request);
+
+        Assert.Equal(exceptionMessage, response.ErrorMessage);
+        Assert.Equal(ResponseStatus.Error, response.ResponseStatus);
+    }
+
+    [Fact, Obsolete("Obsolete")]
     public async Task Can_Handle_Exception_Thrown_By_OnBeforeDeserialization_Handler() {
         const string exceptionMessage = "Thrown from OnBeforeDeserialization";
 
@@ -36,7 +34,7 @@ public class AsyncTests {
     [Fact]
     public async Task Can_Perform_ExecuteGetAsync_With_Response_Type() {
         var request  = new RestRequest("success");
-        var response = await _client.ExecuteAsync<Response>(request);
+        var response = await _client.ExecuteAsync<SuccessResponse>(request);
 
         response.StatusCode.Should().Be(HttpStatusCode.OK);
         response.Data!.Message.Should().Be("Works!");
@@ -57,7 +55,7 @@ public class AsyncTests {
         var request = new RestRequest("timeout").AddBody("Body_Content");
 
         // Half the value of ResponseHandler.Timeout
-        request.Timeout = 200;
+        request.Timeout = TimeSpan.FromMilliseconds(200);
 
         var response = await _client.ExecuteAsync(request);
 
@@ -67,7 +65,7 @@ public class AsyncTests {
     [Fact]
     public async Task Can_Perform_Delete_With_Response_Type() {
         var request  = new RestRequest("delete");
-        var response = await _client.ExecuteAsync<Response>(request, Method.Delete);
+        var response = await _client.ExecuteAsync<SuccessResponse>(request, Method.Delete);
 
         response.StatusCode.Should().Be(HttpStatusCode.OK);
         response.Data!.Message.Should().Be("Works!");
@@ -76,8 +74,15 @@ public class AsyncTests {
     [Fact]
     public async Task Can_Delete_With_Response_Type_using_extension() {
         var request  = new RestRequest("delete");
-        var response = await _client.DeleteAsync<Response>(request);
+        var response = await _client.DeleteAsync<SuccessResponse>(request);
 
         response!.Message.Should().Be("Works!");
     }
+
+    class ThrowingInterceptor(string errorMessage) : Interceptors.Interceptor {
+        public override ValueTask BeforeDeserialization(RestResponse response, CancellationToken cancellationToken)
+            => throw new Exception(errorMessage);
+    }
+
+    public void Dispose() => _client.Dispose();
 }
