@@ -14,10 +14,8 @@
 
 namespace RestSharp.Authenticators;
 
-public abstract class AuthenticatorBase : IAuthenticator {
-    protected AuthenticatorBase(string token) => Token = token;
-
-    protected string Token { get; set; }
+public abstract class AuthenticatorBase(string token) : IAuthenticator {
+    protected string Token { get; set; } = token;
 
     protected abstract ValueTask<Parameter> GetAuthenticationParameter(string accessToken);
 
