@@ -20,8 +20,8 @@ namespace RestSharp;
 [PublicAPI]
 public static partial class RestClientExtensions {
     [PublicAPI]
-    public static RestResponse<T> Deserialize<T>(this IRestClient client, RestResponse response)
-        => client.Serializers.Deserialize<T>(response.Request, response, client.Options);
+    public static ValueTask<RestResponse<T>> Deserialize<T>(this IRestClient client, RestResponse response, CancellationToken cancellationToken)
+        => client.Serializers.Deserialize<T>(response.Request, response, client.Options, cancellationToken);
 
     /// <summary>
     /// Executes the request asynchronously, authenticating if needed
@@ -35,22 +35,10 @@ public static partial class RestClientExtensions {
         RestRequest       request,
         CancellationToken cancellationToken = default
     ) {
-        if (request == null) throw new ArgumentNullException(nameof(request));
+        Ensure.NotNull(request, nameof(request));
 
         var response = await client.ExecuteAsync(request, cancellationToken).ConfigureAwait(false);
-        await OnBeforeDeserialization(response, client.Options).ConfigureAwait(false);
-        return client.Serializers.Deserialize<T>(request, response, client.Options);
-    }
-    
-    /// <summary>
-    /// Will be called before the Data will be serialized
-    /// </summary>
-    /// <param name="raw">RestResponse with Data still in Content</param>
-    /// <param name="options">RestClient options but readonly</param>
-    static async Task OnBeforeDeserialization(RestResponse raw, ReadOnlyRestClientOptions options) {
-        foreach (var interceptor in options.Interceptors) {
-            await interceptor.InterceptBeforeDeserialize(raw);
-        }
+        return await client.Serializers.Deserialize<T>(request, response, client.Options, cancellationToken);
     }
 
     /// <summary>
