@@ -1,25 +1,16 @@
-using System.Net;
-using RestSharp.Tests.Integrated.Server;
+namespace RestSharp.Tests.Integrated;
 
-namespace RestSharp.Tests.Integrated; 
-
-[Collection(nameof(TestServerCollection))]
-public class HttpClientTests {
-    readonly TestServerFixture _fixture;
-
-    public HttpClientTests(TestServerFixture fixture) => _fixture = fixture;
-
+public sealed class HttpClientTests(WireMockTestServer server) : IClassFixture<WireMockTestServer> {
     [Fact]
     public async Task ShouldUseBaseAddress() {
-        using var httpClient = new HttpClient { BaseAddress = _fixture.Server.Url };
-        using var client     = new RestClient(httpClient);
-        
+        using var httpClient = new HttpClient();
+        httpClient.BaseAddress = new Uri(server.Url!);
+        using var client = new RestClient(httpClient);
+
         var request  = new RestRequest("success");
-        var response = await client.ExecuteAsync<Response>(request);
+        var response = await client.ExecuteAsync<SuccessResponse>(request);
 
         response.StatusCode.Should().Be(HttpStatusCode.OK);
         response.Data!.Message.Should().Be("Works!");
     }
-
-    record Response(string Message);
 }
