@@ -1,9 +1,7 @@
-﻿using System.Net;
-using HttpTracer;
+﻿using HttpTracer;
 using RestSharp.Tests.Integrated.Fixtures;
+using RestSharp.Tests.Shared.Extensions;
 using RestSharp.Tests.Shared.Fixtures;
-#pragma warning disable CS8618
-#pragma warning disable CS8601
 
 namespace RestSharp.Tests.Integrated;
 
@@ -12,15 +10,19 @@ public sealed class MultipartFormDataTests : IDisposable {
 
     public MultipartFormDataTests(ITestOutputHelper output) {
         _output = output;
-        _server = SimpleServer.Create(RequestHandler.Handle);
+        _server = WireMockServer.Start();
 
-        var options = new RestClientOptions(_server.Url) {
+        _capturer = _server.ConfigureBodyCapturer(Method.Post);
+        var options = new RestClientOptions($"{_server.Url!}{RequestBodyCapturer.Resource}") {
             ConfigureMessageHandler = handler => new HttpTracerHandler(handler, new OutputLogger(output), HttpMessageParts.All)
         };
         _client = new RestClient(options);
     }
 
-    public void Dispose() => _server.Dispose();
+    public void Dispose() {
+        _server.Dispose();
+        _client.Dispose();
+    }
 
     const string LineBreak = "\r\n";
 
@@ -34,30 +36,22 @@ public sealed class MultipartFormDataTests : IDisposable {
         $"--{{0}}--{LineBreak}";
 
     const string ExpectedFileAndBodyRequestContent =
-        "--{0}"                                                                                                  +
-        $"{LineBreak}{KnownHeaders.ContentType}: application/octet-stream"                                       +
+        "--{0}" +
+        $"{LineBreak}{KnownHeaders.ContentType}: application/octet-stream" +
         $"{LineBreak}{KnownHeaders.ContentDisposition}: form-data; name=\"fileName\"; filename=\"TestFile.txt\"" +
-        $"{LineBreak}{LineBreak}This is a test file for RestSharp.{LineBreak}"                                   +
-        $"--{{0}}{LineBreak}{KnownHeaders.ContentType}: application/json; {CharsetString}"                       +
-        $"{LineBreak}{KnownHeaders.ContentDisposition}: form-data; name=controlName"                             +
-        $"{LineBreak}{LineBreak}test{LineBreak}"                                                                 +
+        $"{LineBreak}{LineBreak}This is a test file for RestSharp.{LineBreak}" +
+        $"--{{0}}{LineBreak}{KnownHeaders.ContentType}: application/json; {CharsetString}" +
+        $"{LineBreak}{KnownHeaders.ContentDisposition}: form-data; name=controlName" +
+        $"{LineBreak}{LineBreak}test{LineBreak}" +
         $"--{{0}}--{LineBreak}";
 
     const string ExpectedDefaultMultipartContentType = "multipart/form-data; boundary=\"{0}\"";
 
     const string ExpectedCustomMultipartContentType = "multipart/vnd.resteasy+form-data; boundary=\"{0}\"";
 
-    readonly SimpleServer _server;
-    readonly RestClient   _client;
-
-    static class RequestHandler {
-        public static string CapturedContentType { get; set; }
-
-        public static void Handle(HttpListenerContext context) {
-            CapturedContentType = context.Request.ContentType;
-            Handlers.Echo(context);
-        }
-    }
+    readonly WireMockServer      _server;
+    readonly RestClient          _client;
+    readonly RequestBodyCapturer _capturer;
 
     static void AddParameters(RestRequest request) {
         request.AddParameter("foo", "bar");
@@ -85,12 +79,12 @@ public sealed class MultipartFormDataTests : IDisposable {
         AddParameters(request);
         request.MultipartFormQuoteBoundary = false;
 
-        var response = await _client.ExecuteAsync(request);
+        await _client.ExecuteAsync(request);
 
         var expected = string.Format(Expected, request.FormBoundary);
 
-        response.Content.Should().Be(expected);
-        RequestHandler.CapturedContentType.Should().Be($"multipart/form-data; boundary={request.FormBoundary}");
+        _capturer.Body.Should().Be(expected);
+        _capturer.ContentType.Should().Be($"multipart/form-data; boundary={request.FormBoundary}");
     }
 
     [Fact]
@@ -106,8 +100,8 @@ public sealed class MultipartFormDataTests : IDisposable {
         _output.WriteLine($"Expected: {expected}");
         _output.WriteLine($"Actual: {response.Content}");
 
-        response.Content.Should().Be(expected);
-        RequestHandler.CapturedContentType.Should().Be($"multipart/form-data; boundary=\"{request.FormBoundary}\"");
+        _capturer.Body.Should().Be(expected);
+        _capturer.ContentType.Should().Be($"multipart/form-data; boundary=\"{request.FormBoundary}\"");
     }
 
     [Fact]
@@ -117,7 +111,7 @@ public sealed class MultipartFormDataTests : IDisposable {
         var path = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Assets", "TestFile.txt");
         request.AddFile("fileName", path);
 
-        request.AddParameter(new BodyParameter("controlName", "test", "application/json"));
+        request.AddParameter(new BodyParameter("controlName", "test", ContentType.Json));
 
         var response = await _client.ExecuteAsync(request);
 
@@ -129,8 +123,8 @@ public sealed class MultipartFormDataTests : IDisposable {
         _output.WriteLine($"Expected: {expectedFileAndBodyRequestContent}");
         _output.WriteLine($"Actual: {response.Content}");
 
-        response.Content.Should().Be(expectedFileAndBodyRequestContent);
-        expectedDefaultMultipartContentType.Should().Be(RequestHandler.CapturedContentType);
+        _capturer.Body.Should().Be(expectedFileAndBodyRequestContent);
+        _capturer.ContentType.Should().Be(expectedDefaultMultipartContentType);
     }
 
     [Fact]
@@ -142,16 +136,16 @@ public sealed class MultipartFormDataTests : IDisposable {
         const string customContentType = "multipart/vnd.resteasy+form-data";
         request.AddHeader(KnownHeaders.ContentType, customContentType);
         request.AddFile("fileName", path);
-        request.AddParameter(new BodyParameter("controlName", "test", "application/json"));
+        request.AddParameter(new BodyParameter("controlName", "test", ContentType.Json));
 
-        var response = await _client.ExecuteAsync(request);
+        await _client.ExecuteAsync(request);
         var boundary = request.FormBoundary;
 
         var expectedFileAndBodyRequestContent  = string.Format(ExpectedFileAndBodyRequestContent, boundary);
         var expectedCustomMultipartContentType = string.Format(ExpectedCustomMultipartContentType, boundary);
 
-        response.Content.Should().Be(expectedFileAndBodyRequestContent);
-        RequestHandler.CapturedContentType.Should().Be(expectedCustomMultipartContentType);
+        _capturer.Body.Should().Be(expectedFileAndBodyRequestContent);
+        _capturer.ContentType.Should().Be(expectedCustomMultipartContentType);
     }
 
     [Fact]
@@ -163,14 +157,14 @@ public sealed class MultipartFormDataTests : IDisposable {
         var path = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Assets", "TestFile.txt");
         request.AddFile("fileName", path);
 
-        request.AddParameter(new BodyParameter("controlName", "test", "application/json"));
+        request.AddParameter(new BodyParameter("controlName", "test", ContentType.Json));
 
-        var response = await _client.ExecuteAsync(request);
+        await _client.ExecuteAsync(request);
         var boundary = request.FormBoundary;
 
         var expectedFileAndBodyRequestContent = string.Format(ExpectedFileAndBodyRequestContent, boundary);
 
-        response.Content.Should().Be(expectedFileAndBodyRequestContent);
+        _capturer.Body.Should().Be(expectedFileAndBodyRequestContent);
     }
 
     [Fact]
@@ -179,12 +173,37 @@ public sealed class MultipartFormDataTests : IDisposable {
 
         AddParameters(request);
 
-        var response = await _client.ExecuteAsync(request);
+        await _client.ExecuteAsync(request);
 
         var boundary = request.FormBoundary;
         var expected = string.Format(Expected, boundary);
 
-        response.Content.Should().Be(expected);
+        _capturer.Body.Should().Be(expected);
     }
+    
+    [Fact]
+    public async Task MultipartFormData_Without_File_Creates_A_Valid_RequestBody() {
+        using var client = new RestClient(_server.Url!);
 
+        var request = new RestRequest(RequestBodyCapturer.Resource, Method.Post) {
+            AlwaysMultipartFormData = true,
+        };
+        var capturer = _server.ConfigureBodyCapturer(Method.Post);
+
+        const string bodyData      = "abc123 foo bar baz BING!";
+        const string multipartName = "mybody";
+
+        request.AddParameter(new BodyParameter(multipartName, bodyData, ContentType.Plain));
+
+        await client.ExecuteAsync(request);
+
+        var expectedBody = new[] {
+            ContentTypeString,
+            $"{ContentDispositionString} name={multipartName}",
+            bodyData
+        };
+
+        var actual = capturer.Body!.Replace("\n", string.Empty).Split('\r');
+        actual.Should().Contain(expectedBody);
+    }
 }
