@@ -1,26 +1,20 @@
-﻿using System.Net;
-using RestSharp.Tests.Shared.Fixtures;
-
-namespace RestSharp.Tests.Integrated;
+﻿namespace RestSharp.Tests.Integrated;
 
 public sealed class NonProtocolExceptionHandlingTests : IDisposable {
+    public NonProtocolExceptionHandlingTests()
+        => _server
+            .Given(Request.Create().WithPath("/timeout"))
+            .RespondWith(Response.Create().WithDelay(TimeSpan.FromSeconds(1)));
+
     // ReSharper disable once ClassNeverInstantiated.Local
     class StupidClass {
         // ReSharper disable once UnusedMember.Local
         public string Property { get; set; } = null!;
     }
 
-    /// <summary>
-    /// Simulates a long server process that should result in a client timeout
-    /// </summary>
-    /// <param name="context"></param>
-    static void TimeoutHandler(HttpListenerContext context) => Thread.Sleep(101000);
-
-    public NonProtocolExceptionHandlingTests() => _server = SimpleServer.Create(TimeoutHandler);
-
     public void Dispose() => _server.Dispose();
 
-    readonly SimpleServer _server;
+    readonly WireMockServer _server = WireMockServer.Start();
 
     /// <summary>
     /// Success of this test is based largely on the behavior of your current DNS.
@@ -28,18 +22,19 @@ public sealed class NonProtocolExceptionHandlingTests : IDisposable {
     /// </summary>
     [Fact]
     public async Task Handles_Non_Existent_Domain() {
-        var client   = new RestClient("http://nonexistantdomainimguessing.org");
+        using var client = new RestClient("http://nonexistantdomainimguessing.org");
+
         var request  = new RestRequest("foo");
         var response = await client.ExecuteAsync(request);
 
-        Assert.Equal(ResponseStatus.Error, response.ResponseStatus);
+        response.ResponseStatus.Should().Be(ResponseStatus.Error);
     }
 
     [Fact]
     public async Task Handles_HttpClient_Timeout_Error() {
-        var client = new RestClient(new HttpClient {Timeout = TimeSpan.FromMilliseconds(500)});
+        using var client = new RestClient(new HttpClient { Timeout = TimeSpan.FromMilliseconds(500) });
 
-        var request = new RestRequest($"{_server.Url}/404");
+        var request  = new RestRequest($"{_server.Url}/timeout");
         var response = await client.ExecuteAsync(request);
 
         response.ErrorException.Should().BeOfType<TaskCanceledException>();
@@ -48,9 +43,9 @@ public sealed class NonProtocolExceptionHandlingTests : IDisposable {
 
     [Fact]
     public async Task Handles_Server_Timeout_Error() {
-        var client = new RestClient(_server.Url);
+        using var client = new RestClient(_server.Url!);
 
-        var request = new RestRequest("404") { Timeout = 500 };
+        var request  = new RestRequest("404") { Timeout = TimeSpan.FromMilliseconds(500) };
         var response = await client.ExecuteAsync(request);
 
         response.ErrorException.Should().BeOfType<TaskCanceledException>();
@@ -59,9 +54,10 @@ public sealed class NonProtocolExceptionHandlingTests : IDisposable {
 
     [Fact]
     public async Task Handles_Server_Timeout_Error_With_Deserializer() {
-        var client   = new RestClient(_server.Url);
-        var request  = new RestRequest("404") { Timeout = 500 };
-        var response = await client.ExecuteAsync<TestResponse>(request);
+        using var client = new RestClient(_server.Url!);
+
+        var request  = new RestRequest("timeout") { Timeout = TimeSpan.FromMilliseconds(500) };
+        var response = await client.ExecuteAsync<SuccessResponse>(request);
 
         response.Data.Should().BeNull();
         response.ErrorException.Should().BeOfType<TaskCanceledException>();
@@ -70,7 +66,7 @@ public sealed class NonProtocolExceptionHandlingTests : IDisposable {
 
     [Fact]
     public async Task Task_Handles_Non_Existent_Domain() {
-        var client = new RestClient("http://this.cannot.exist:8001");
+        using var client = new RestClient("http://this.cannot.exist:8001");
 
         var request = new RestRequest("/") {
             RequestFormat = DataFormat.Json,
