@@ -1,12 +1,7 @@
-using System.Net;
-using RestSharp.Tests.Shared.Fixtures;
-
 namespace RestSharp.Tests.Integrated; 
 
 public sealed class ResourceStringParametersTests : IDisposable {
-    readonly SimpleServer _server;
-
-    public ResourceStringParametersTests() => _server = SimpleServer.Create(RequestHandler.Handle);
+    readonly WireMockServer _server = WireMockServer.Start();
 
     public void Dispose() => _server.Dispose();
 
@@ -14,23 +9,20 @@ public sealed class ResourceStringParametersTests : IDisposable {
     public async Task Should_keep_to_parameters_with_the_same_name() {
         const string parameters = "?priority=Low&priority=Medium";
 
-        var client  = new RestClient(_server.Url);
+        var url = "";
+        _server
+            .Given(Request.Create())
+            .RespondWith(Response.Create().WithCallback(req => {
+                url = req.Url;
+                return new ResponseMessage();
+            }));
+        
+        using var client  = new RestClient(_server.Url!);
         var request = new RestRequest(parameters);
 
         await client.GetAsync(request);
 
-        var query = RequestHandler.Url?.Query;
+        var query = new Uri(url).Query;
         query.Should().Be(parameters);
     }
-
-    #nullable disable
-    static class RequestHandler {
-        public static Uri Url { get; private set; }
-
-        public static void Handle(HttpListenerContext context) {
-            Url = context.Request.Url;
-            Handlers.Echo(context);
-        }
-    }
-    #nullable enable
 }
