@@ -1,16 +1,26 @@
-﻿using System.Net;
-using RestSharp.Serializers.Xml;
-using RestSharp.Tests.Shared.Extensions;
-using RestSharp.Tests.Shared.Fixtures;
+﻿using RestSharp.Serializers.Xml;
 
 namespace RestSharp.Tests.Integrated; 
 
 public class RootElementTests {
     [Fact]
     public async Task Copy_RootElement_From_Request_To_IWithRootElement_Deserializer() {
-        using var server = HttpServerFixture.StartServer("success", Handle);
+        using var server = WireMockServer.Start();
 
-        var client = new RestClient(server.Url, configureSerialization: cfg => cfg.UseXmlSerializer());
+        const string xmlBody =
+            """
+            <?xml version="1.0" encoding="utf-8" ?>
+            <Response>
+                <Success>
+                    <Message>Works!</Message>
+                </Success>
+            </Response>
+            """;
+        server
+            .Given(Request.Create().WithPath("/success"))
+            .RespondWith(Response.Create().WithBody(xmlBody).WithHeader(KnownHeaders.ContentType, ContentType.Xml));
+
+        using var client = new RestClient(server.Url!, configureSerialization: cfg => cfg.UseXmlSerializer());
 
         var request = new RestRequest("success") { RootElement = "Success" };
 
@@ -18,19 +28,5 @@ public class RootElementTests {
 
         response.Data.Should().NotBeNull();
         response.Data!.Message.Should().Be("Works!");
-
-        static void Handle(HttpListenerRequest req, HttpListenerResponse response) {
-            response.StatusCode = 200;
-            response.Headers.Add(KnownHeaders.ContentType, ContentType.Xml);
-
-            response.OutputStream.WriteStringUtf8(
-                @"<?xml version=""1.0"" encoding=""utf-8"" ?>
-<Response>
-    <Success>
-        <Message>Works!</Message>
-    </Success>
-</Response>"
-            );
-        }
     }
 }
