@@ -1,7 +1,4 @@
-﻿using System.Net;
-using RestSharp.Serializers.Xml;
-using RestSharp.Tests.Shared.Extensions;
-using RestSharp.Tests.Shared.Fixtures;
+﻿using RestSharp.Serializers.Xml;
 
 // ReSharper disable UnusedMember.Local
 // ReSharper disable InconsistentNaming
@@ -10,71 +7,28 @@ namespace RestSharp.Tests.Integrated;
 
 public sealed class StatusCodeTests : IDisposable {
     public StatusCodeTests() {
-        _server = SimpleServer.Create(UrlToStatusCodeHandler);
-        _client = new RestClient(_server.Url, configureSerialization: cfg => cfg.UseXmlSerializer());
+        _client = new RestClient(_server.Url!, configureSerialization: cfg => cfg.UseXmlSerializer());
+        _server
+            .Given(Request.Create())
+            .RespondWith(Response.Create().WithCallback(CreateResponse));
+        return;
+
+        ResponseMessage CreateResponse(IRequestMessage request) {
+            var url = new Uri(request.Url);
+
+            return new ResponseMessage() {
+                StatusCode = int.Parse(url.Segments.Last())
+            };
+        }
     }
 
-    public void Dispose() => _server.Dispose();
-
-    readonly SimpleServer _server;
-    readonly RestClient   _client;
-
-    static void UrlToStatusCodeHandler(HttpListenerContext obj) => obj.Response.StatusCode = int.Parse(obj.Request.Url!.Segments.Last());
-
-    [Fact]
-    public async Task ContentType_Additional_Information() {
-        _server.SetHandler(Handlers.Generic<ResponseHandler>());
-
-        var request = new RestRequest("", Method.Post) {
-            RequestFormat = DataFormat.Json,
-            Resource      = "contenttype_odata"
-        };
-        request.AddBody("bodyadsodajjd");
-        request.AddHeader("X-RequestDigest", "xrequestdigestasdasd");
-        request.AddHeader(KnownHeaders.Accept, $"{ContentType.Json}; odata=verbose");
-        request.AddHeader(KnownHeaders.ContentType, $"{ContentType.Json}; odata=verbose");
-
-        var response = await _client.ExecuteAsync<TestResponse>(request);
-
-        response.StatusCode.Should().Be(HttpStatusCode.OK);
-        response.IsSuccessful.Should().BeTrue();
-        response.IsSuccessStatusCode.Should().BeTrue();
+    public void Dispose() {
+        _server.Dispose();
+        _client.Dispose();
     }
 
-    [Fact]
-    public async Task Handles_Default_Root_Element_On_No_Error() {
-        _server.SetHandler(Handlers.Generic<ResponseHandler>());
-
-        var request = new RestRequest("success") {
-            RootElement = "Success"
-        };
-
-        request.OnBeforeDeserialization = resp => {
-            if (resp.StatusCode == HttpStatusCode.NotFound) request.RootElement = "Error";
-        };
-
-        var response = await _client.ExecuteAsync<TestResponse>(request);
-
-        response.StatusCode.Should().Be(HttpStatusCode.OK);
-        response.Data!.Message.Should().Be("Works!");
-    }
-
-    [Fact]
-    public async Task Handles_Different_Root_Element_On_Http_Error() {
-        _server.SetHandler(Handlers.Generic<ResponseHandler>());
-
-        var request = new RestRequest("error") {
-            RootElement = "Success",
-            OnBeforeDeserialization = resp => {
-                if (resp.StatusCode == HttpStatusCode.BadRequest) resp.RootElement = "Error";
-            }
-        };
-
-        var response = await _client.ExecuteAsync<TestResponse>(request);
-
-        response.StatusCode.Should().Be(HttpStatusCode.BadRequest);
-        response.Data!.Message.Should().Be("Not found!");
-    }
+    readonly WireMockServer _server = WireMockServer.Start();
+    readonly RestClient     _client;
 
     [Fact]
     public async Task Handles_GET_Request_404_Error() {
@@ -128,54 +82,4 @@ public sealed class StatusCodeTests : IDisposable {
         response.IsSuccessful.Should().BeFalse();
         response.IsSuccessStatusCode.Should().BeFalse();
     }
-}
-
-public class ResponseHandler {
-    void contenttype_odata(HttpListenerContext context) {
-        var contentType      = context.Request.Headers[KnownHeaders.ContentType];
-        var hasCorrectHeader = contentType!.Contains($"{ContentType.Json}; odata=verbose");
-        context.Response.StatusCode = hasCorrectHeader ? 200 : 400;
-    }
-
-    void error(HttpListenerContext context) {
-        context.Response.StatusCode = 400;
-        context.Response.Headers.Add(KnownHeaders.ContentType, ContentType.Xml);
-
-        context.Response.OutputStream.WriteStringUtf8(
-            @"<?xml version=""1.0"" encoding=""utf-8"" ?>
-<Response>
-    <Error>
-        <Message>Not found!</Message>
-    </Error>
-</Response>"
-        );
-    }
-
-    void errorwithbody(HttpListenerContext context) {
-        context.Response.StatusCode = 400;
-        context.Response.Headers.Add(KnownHeaders.ContentType, "application/xml");
-
-        context.Response.OutputStream.WriteStringUtf8(
-            @"<?xml version=""1.0"" encoding=""utf-8"" ?>
-<Response>
-    <Error>
-        <Message>Not found!</Message>
-    </Error>
-</Response>"
-        );
-    }
-
-    void success(HttpListenerContext context)
-        => context.Response.OutputStream.WriteStringUtf8(
-            @"<?xml version=""1.0"" encoding=""utf-8"" ?>
-<Response>
-    <Success>
-        <Message>Works!</Message>
-    </Success>
-</Response>"
-        );
-}
-
-public class TestResponse {
-    public string Message { get; set; } = null!;
 }
