@@ -1,15 +1,14 @@
-﻿using System.Net;
+﻿using System.Text;
 using RestSharp.Serializers.Xml;
-using RestSharp.Tests.Shared.Extensions;
-using RestSharp.Tests.Shared.Fixtures;
+using WireMock.Types;
+using WireMock.Util;
 
 // ReSharper disable UnusedAutoPropertyAccessor.Local
 
 namespace RestSharp.Tests.Integrated;
 
 public sealed class StructuredSyntaxSuffixTests : IDisposable {
-    readonly TestHttpServer _server;
-    readonly string         _url;
+    readonly WireMockServer _server;
 
     class Person {
         public string Name { get; set; } = null!;
@@ -18,16 +17,26 @@ public sealed class StructuredSyntaxSuffixTests : IDisposable {
     }
 
     const string XmlContent  = "<Person><name>Bob</name><age>50</age></Person>";
-    const string JsonContent = @"{ ""name"":""Bob"", ""age"":50 }";
+    const string JsonContent = """{ "name":"Bob", "age":50 }""";
 
     public StructuredSyntaxSuffixTests() {
-        _server = new TestHttpServer(0, "", HandleRequest);
-        _url    = $"http://localhost:{_server.Port}";
+        _server = WireMockServer.Start();
+        _server.Given(Request.Create().WithPath("/").UsingGet()).RespondWith(Response.Create().WithCallback(Handle));
+        return;
 
-        static void HandleRequest(HttpListenerRequest request, HttpListenerResponse response, Dictionary<string, string> p) {
-            response.ContentType = request.QueryString["ct"];
-            response.OutputStream.WriteStringUtf8(request.QueryString["c"]);
-            response.StatusCode = 200;
+        static ResponseMessage Handle(IRequestMessage request) {
+            var response = new ResponseMessage {
+                Headers = new Dictionary<string, WireMockList<string>> {
+                    [KnownHeaders.ContentType] = new(request.Query!["ct"])
+                },
+                StatusCode   = 200,
+                BodyData = new BodyData {
+                    BodyAsString = request.Query["c"].First(),
+                    Encoding = Encoding.UTF8,
+                    DetectedBodyType = BodyType.String
+                }
+            };
+            return response;
         }
     }
 
@@ -35,10 +44,10 @@ public sealed class StructuredSyntaxSuffixTests : IDisposable {
 
     [Fact]
     public async Task By_default_application_json_content_type_should_deserialize_as_JSON() {
-        var client = new RestClient(_url);
+        using var client = new RestClient(_server.Url!);
 
         var request = new RestRequest()
-            .AddParameter("ct", "application/json")
+            .AddParameter("ct", ContentType.Json)
             .AddParameter("c", JsonContent);
 
         var response = await client.ExecuteAsync<Person>(request);
@@ -49,7 +58,7 @@ public sealed class StructuredSyntaxSuffixTests : IDisposable {
 
     [Fact]
     public async Task By_default_content_types_with_JSON_structured_syntax_suffix_should_deserialize_as_JSON() {
-        var client = new RestClient(_url);
+        using var client = new RestClient(_server.Url!);
 
         var request = new RestRequest()
             .AddParameter("ct", "application/vnd.somebody.something+json")
@@ -63,7 +72,7 @@ public sealed class StructuredSyntaxSuffixTests : IDisposable {
 
     [Fact]
     public async Task By_default_content_types_with_XML_structured_syntax_suffix_should_deserialize_as_XML() {
-        var client = new RestClient(_url, configureSerialization: cfg => cfg.UseXmlSerializer());
+        using var client = new RestClient(_server.Url!, configureSerialization: cfg => cfg.UseXmlSerializer());
 
         var request = new RestRequest()
             .AddParameter("ct", "application/vnd.somebody.something+xml")
@@ -77,7 +86,7 @@ public sealed class StructuredSyntaxSuffixTests : IDisposable {
 
     [Fact]
     public async Task By_default_text_xml_content_type_should_deserialize_as_XML() {
-        var client = new RestClient(_url, configureSerialization: cfg => cfg.UseXmlSerializer());
+        using var client = new RestClient(_server.Url!, configureSerialization: cfg => cfg.UseXmlSerializer());
 
         var request = new RestRequest()
             .AddParameter("ct", "text/xml")
