@@ -1,24 +1,30 @@
-using System.Net;
-using RestSharp.Tests.Integrated.Server;
+// ReSharper disable MethodHasAsyncOverload
+
+using HttpMultipartParser;
+using RestSharp.Extensions;
+using RestSharp.Tests.Shared.Fixtures;
 
 namespace RestSharp.Tests.Integrated;
 
-[Collection(nameof(TestServerCollection))]
-public class UploadFileTests {
+public sealed class UploadFileTests : IDisposable {
     readonly ITestOutputHelper _output;
     readonly RestClient        _client;
     readonly string            _basePath = AppDomain.CurrentDomain.BaseDirectory;
     readonly string            _path;
     readonly UploadResponse    _expected;
+    readonly WireMockServer    _server = WireMockServer.Start();
 
     const string Filename = "Koala.jpg";
 
-    public UploadFileTests(TestServerFixture fixture, ITestOutputHelper output) {
-        _output = output;
-        // _client = new RestClient(new RestClientOptions(fixture.Server.Url) { ThrowOnAnyError = true });
-        _client   = new RestClient(new RestClientOptions(fixture.Server.Url) { ThrowOnAnyError = false });
+    public UploadFileTests(ITestOutputHelper output) {
+        _output   = output;
+        _client   = new RestClient(new RestClientOptions(_server.Url!));
         _path     = Path.Combine(_basePath, "Assets", Filename);
         _expected = new UploadResponse(Filename, new FileInfo(_path).Length, true);
+
+        _server
+            .Given(Request.Create().WithPath("/upload"))
+            .RespondWith(Response.Create().WithCallback(HandleUpload));
     }
 
     [Fact]
@@ -34,7 +40,7 @@ public class UploadFileTests {
 
     [Fact]
     public async Task Should_upload_from_bytes() {
-        var bytes = await File.ReadAllBytesAsync(_path);
+        var bytes = File.ReadAllBytes(_path);
 
         var request  = new RestRequest("upload").AddFile("file", bytes, Filename);
         var response = await _client.ExecutePostAsync<UploadResponse>(request);
@@ -52,11 +58,12 @@ public class UploadFileTests {
         response.Data.Should().BeEquivalentTo(_expected);
     }
 
+    // This test fails because MultipartFormDataParser doesn't understand filename*
     [Fact]
     public async Task Should_upload_from_stream_non_ascii() {
         const string nonAsciiFilename = "Präsentation_Export.zip";
 
-        var options = new FileParameterOptions { DisableFilenameEncoding = true, DisableFilenameStar = false};
+        var options = new FileParameterOptions { DisableFilenameEncoding = true, DisableFilenameStar = false };
 
         var request = new RestRequest("upload")
             .AddFile("file", () => File.OpenRead(_path), nonAsciiFilename, options: options)
@@ -65,5 +72,44 @@ public class UploadFileTests {
 
         _output.WriteLine(response.Content);
         response.Data.Should().BeEquivalentTo(new UploadResponse(nonAsciiFilename, new FileInfo(_path).Length, true));
+    }
+
+    static async Task<ResponseMessage> HandleUpload(IRequestMessage request) {
+        var response = new ResponseMessage();
+
+        var checkFile = request.Query == null ||
+            request.Query.Count == 0 ||
+            request.Query.ContainsKey("checkFile") && bool.Parse(request.Query["checkFile"][0]);
+
+        using var stream = new MemoryStream(request.BodyAsBytes!);
+        var       form   = await MultipartFormDataParser.ParseAsync(stream);
+        if (form.Files.Count == 0) return response;
+
+        var fileSection = form.Files[0];
+        var fileLength  = fileSection.Data.Length;
+        var fileName = fileSection.FileName;
+
+        // ReSharper disable once InvertIf
+        if (checkFile) {
+            var assetPath = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Assets");
+
+            try {
+                var expected = File.ReadAllBytes(Path.Combine(assetPath, fileName));
+                fileSection.Data.Seek(0, SeekOrigin.Begin);
+                var received = await fileSection.Data.ReadAsBytes(default);
+                var equal    = received.SequenceEqual(expected);
+                return WireMockTestServer.CreateJson(new UploadResponse(fileName, fileLength, equal));
+            }
+            catch (Exception) {
+                return response;
+            }
+        }
+
+        return WireMockTestServer.CreateJson(new UploadResponse(fileName, fileLength, true));
+    }
+
+    public void Dispose() {
+        _client.Dispose();
+        _server.Dispose();
     }
 }
