@@ -1,96 +1,69 @@
-using System.Net;
-using System.Text;
 using RestSharp.Serializers.NewtonsoftJson;
 using RestSharp.Tests.Shared.Extensions;
 using RestSharp.Tests.Shared.Fixtures;
 
 namespace RestSharp.Tests.Serializers.Json.NewtonsoftJson;
 
-public class IntegratedSimpleTests {
-    string _body;
-
-    void CaptureBody(HttpListenerRequest request, HttpListenerResponse response) => _body = request.InputStream.StreamToString();
-
+public sealed class IntegratedSimpleTests : IDisposable {
     static readonly Fixture Fixture = new();
 
+    readonly WireMockServer _server = WireMockServer.Start();
+    readonly RestClient     _client;
+
+    public IntegratedSimpleTests() => _client = new RestClient(_server.Url!, configureSerialization: cfg => cfg.UseNewtonsoftJson());
+
     [Fact]
-    public async Task Use_JsonNet_For_Requests() {
-        using var server = HttpServerFixture.StartServer(CaptureBody);
-        _body = null;
+    public async Task Should_serialize_request() {
+        var capturer   = _server.ConfigureBodyCapturer(Method.Post, false);
         var serializer = new JsonNetSerializer();
+        var testData   = Fixture.Create<TestClass>();
+        var request    = new RestRequest().AddJsonBody(testData);
 
-        var testData = Fixture.Create<TestClass>();
-
-        var client  = new RestClient(server.Url, configureSerialization: cfg => cfg.UseNewtonsoftJson());
-        var request = new RestRequest().AddJsonBody(testData);
-
-        await client.PostAsync(request);
-
-        var actual = serializer.Deserialize<TestClass>(new RestResponse(request) { Content = _body! });
+        await _client.PostAsync(request);
+        var actual = serializer.Deserialize<TestClass>(new RestResponse(request) { Content = capturer.Body! });
 
         actual.Should().BeEquivalentTo(testData);
     }
 
     [Fact]
-    public async Task Use_JsonNet_For_Response() {
+    public async Task Should_deserialize_response() {
         var expected = Fixture.Create<TestClass>();
 
-        using var server = HttpServerFixture.StartServer(
-            (_, response) => {
-                var serializer = new JsonNetSerializer();
+        _server
+            .Given(Request.Create().WithPath("/").UsingGet())
+            .RespondWith(Response.Create().WithBodyAsJson(expected));
 
-                response.ContentType     = "application/json";
-                response.ContentEncoding = Encoding.UTF8;
-                response.OutputStream.WriteStringUtf8(serializer.Serialize(expected)!);
-            }
-        );
-
-        var client = new RestClient(server.Url, configureSerialization: cfg => cfg.UseNewtonsoftJson());
-
-        var actual = await client.GetAsync<TestClass>(new RestRequest());
-
+        var actual = await _client.GetAsync<TestClass>(new RestRequest());
         actual.Should().BeEquivalentTo(expected);
     }
 
     [Fact]
-    public async Task DeserilizationFails_IsSuccessful_Should_BeFalse() {
-        using var server = HttpServerFixture.StartServer(
-            (_, response) => {
-                response.StatusCode      = (int)HttpStatusCode.OK;
-                response.ContentType     = "application/json";
-                response.ContentEncoding = Encoding.UTF8;
-                response.OutputStream.WriteStringUtf8("invalid json");
-            }
-        );
+    public async Task Invalid_json_body_request_should_fail() {
+        _server
+            .Given(Request.Create().WithPath("/").UsingGet())
+            .RespondWith(Response.Create().WithBody("invalid json").WithHeader(KnownHeaders.ContentType, ContentType.Json));
 
-        var client = new RestClient(server.Url, configureSerialization: cfg => cfg.UseNewtonsoftJson());
 
-        var response = await client.ExecuteAsync<TestClass>(new RestRequest());
-
+        var response = await _client.ExecuteAsync<TestClass>(new RestRequest());
         response.IsSuccessStatusCode.Should().BeTrue();
         response.IsSuccessful.Should().BeFalse();
     }
 
     [Fact]
-    public async Task DeserilizationSucceeds_IsSuccessful_Should_BeTrue() {
+    public async Task Valid_json_response_should_succeed() {
         var item = Fixture.Create<TestClass>();
 
-        using var server = HttpServerFixture.StartServer(
-            (_, response) => {
-                var serializer = new JsonNetSerializer();
+        _server
+            .Given(Request.Create().WithPath("/").UsingGet())
+            .RespondWith(Response.Create().WithBodyAsJson(item));
 
-                response.StatusCode      = (int)HttpStatusCode.OK;
-                response.ContentType     = "application/json";
-                response.ContentEncoding = Encoding.UTF8;
-                response.OutputStream.WriteStringUtf8(serializer.Serialize(item)!);
-            }
-        );
-
-        var client = new RestClient(server.Url, configureSerialization: cfg => cfg.UseNewtonsoftJson());
-
-        var response = await client.ExecuteAsync<TestClass>(new RestRequest());
-
+        var response = await _client.ExecuteAsync<TestClass>(new RestRequest());
         response.IsSuccessStatusCode.Should().BeTrue();
         response.IsSuccessful.Should().BeTrue();
+    }
+
+    public void Dispose() {
+        _server?.Dispose();
+        _client.Dispose();
     }
 }
