@@ -1,96 +1,69 @@
-using System.Net;
-using System.Text;
 using RestSharp.Serializers.Json;
 using RestSharp.Tests.Shared.Extensions;
 using RestSharp.Tests.Shared.Fixtures;
 
 namespace RestSharp.Tests.Serializers.Json.SystemTextJson;
 
-public class SystemTextJsonTests {
+public sealed class SystemTextJsonTests : IDisposable {
     static readonly Fixture Fixture = new();
 
-    string _body;
+    readonly WireMockServer _server = WireMockServer.Start();
+    readonly RestClient     _client;
+
+    public SystemTextJsonTests() => _client = new RestClient(_server.Url!);
 
     [Fact]
-    public async Task Use_JsonNet_For_Requests() {
-        using var server = HttpServerFixture.StartServer(CaptureBody);
-        _body = null;
+    public async Task Should_serialize_request() {
         var serializer = new SystemTextJsonSerializer();
+        var capturer   = _server.ConfigureBodyCapturer(Method.Post, false);
 
         var testData = Fixture.Create<TestClass>();
+        var request  = new RestRequest().AddJsonBody(testData);
 
-        var client  = new RestClient(server.Url);
-        var request = new RestRequest().AddJsonBody(testData);
+        await _client.PostAsync(request);
 
-        await client.PostAsync(request);
-
-        var actual = serializer.Deserialize<TestClass>(new RestResponse(request) { Content = _body });
-
+        var actual = serializer.Deserialize<TestClass>(new RestResponse(request) { Content = capturer.Body });
         actual.Should().BeEquivalentTo(testData);
-
-        void CaptureBody(HttpListenerRequest req, HttpListenerResponse response) => _body = req.InputStream.StreamToString();
     }
 
     [Fact]
-    public async Task Use_JsonNet_For_Response() {
+    public async Task Should_deserialize_response_with_GenericGet() {
         var expected = Fixture.Create<TestClass>();
 
-        using var server = HttpServerFixture.StartServer(
-            (_, response) => {
-                var serializer = new SystemTextJsonSerializer();
+        _server
+            .Given(Request.Create().WithPath("/").UsingGet())
+            .RespondWith(Response.Create().WithBodyAsJson(expected));
 
-                response.ContentType     = "application/json";
-                response.ContentEncoding = Encoding.UTF8;
-                response.OutputStream.WriteStringUtf8(serializer.Serialize(expected)!);
-            }
-        );
-
-        var client = new RestClient(server.Url, configureSerialization: cfg => cfg.UseSystemTextJson());
-
-        var actual = await client.GetAsync<TestClass>(new RestRequest());
-
+        var actual = await _client.GetAsync<TestClass>(new RestRequest());
         actual.Should().BeEquivalentTo(expected);
     }
 
     [Fact]
-    public async Task DeserilizationFails_IsSuccessful_Should_BeFalse() {
-        using var server = HttpServerFixture.StartServer(
-            (_, response) => {
-                response.StatusCode      = (int)HttpStatusCode.OK;
-                response.ContentType     = "application/json";
-                response.ContentEncoding = Encoding.UTF8;
-                response.OutputStream.WriteStringUtf8("invalid json");
-            }
-        );
+    public async Task Posting_invalid_json_should_fail() {
+        _server
+            .Given(Request.Create().WithPath("/").UsingGet())
+            .RespondWith(Response.Create().WithBody("invalid json").WithHeader(KnownHeaders.ContentType, ContentType.Json));
 
-        var client = new RestClient(server.Url, configureSerialization: cfg => cfg.UseSystemTextJson());
-
-        var response = await client.ExecuteAsync<TestClass>(new RestRequest());
-
+        var response = await _client.ExecuteAsync<TestClass>(new RestRequest());
         response.IsSuccessStatusCode.Should().BeTrue();
         response.IsSuccessful.Should().BeFalse();
     }
 
     [Fact]
-    public async Task DeserilizationSucceeds_IsSuccessful_Should_BeTrue() {
+    public async Task Receiving_valid_json_should_succeed() {
         var item = Fixture.Create<TestClass>();
 
-        using var server = HttpServerFixture.StartServer(
-            (_, response) => {
-                var serializer = new SystemTextJsonSerializer();
+        _server
+            .Given(Request.Create().WithPath("/").UsingGet())
+            .RespondWith(Response.Create().WithBodyAsJson(item));
 
-                response.StatusCode      = (int)HttpStatusCode.OK;
-                response.ContentType     = "application/json";
-                response.ContentEncoding = Encoding.UTF8;
-                response.OutputStream.WriteStringUtf8(serializer.Serialize(item)!);
-            }
-        );
-
-        var client = new RestClient(server.Url, configureSerialization: cfg => cfg.UseSystemTextJson());
-
-        var response = await client.ExecuteAsync<TestClass>(new RestRequest());
-
+        var response = await _client.ExecuteAsync<TestClass>(new RestRequest());
         response.IsSuccessStatusCode.Should().BeTrue();
         response.IsSuccessful.Should().BeTrue();
+    }
+
+    public void Dispose() {
+        _server?.Dispose();
+        _client.Dispose();
     }
 }
