@@ -3,10 +3,10 @@
 public class AddRangeTests {
     [Fact]
     public async Task ShouldParseOutLongRangeSpecifier() {
-        var        restClient = new RestClient("http://localhost");
-        var        req        = new RestRequest("bob");
-        const long start      = (long)int.MaxValue + 1;
-        const long end        = start + 1;
+        using var restClient = new RestClient("http://localhost");
+        var req = new RestRequest("bob");
+        const long start = (long)int.MaxValue + 1;
+        const long end = start + 1;
 
         req.AddHeader("Range", $"pages={start}-{end}");
         await restClient.ExecuteAsync(req);
@@ -14,8 +14,8 @@ public class AddRangeTests {
 
     [Fact]
     public async Task ShouldParseOutRangeSpecifier() {
-        var restClient = new RestClient("http://localhost");
-        var req        = new RestRequest("bob");
+        using var restClient = new RestClient("http://localhost");
+        var req = new RestRequest("bob");
 
         req.AddHeader("Range", "pages=1-2");
         await restClient.ExecuteAsync(req);
