@@ -1,4 +1,5 @@
 using System.Net;
+using RestSharp.Tests.Fixtures;
 using RichardSzalay.MockHttp;
 
 namespace RestSharp.Tests;
@@ -6,25 +7,13 @@ namespace RestSharp.Tests;
 public class MultipartFormTests {
     [Fact]
     public async Task ShouldHaveJsonContentType() {
-        const string url = "https://dummy.org";
-
         var jsonData = new {
             Company = "Microsoft",
             ZipCode = "LS339",
             Country = "USA"
         };
 
-        var mockHttp = new MockHttpMessageHandler();
-
-        mockHttp
-            .When(HttpMethod.Post, url)
-            .With(CheckRequest)
-            .Respond(HttpStatusCode.OK);
-
-        var options = new RestClientOptions(url) {
-            ConfigureMessageHandler = h => mockHttp
-        };
-        var client = new RestClient(options);
+        using var client = MockHttpClient.Create(Method.Post, req => req.With(CheckRequest).Respond(HttpStatusCode.OK));
 
         var request = new RestRequest {
             Method                  = Method.Post,
@@ -34,11 +23,12 @@ public class MultipartFormTests {
 
         var response = await client.ExecuteAsync(request);
         response.StatusCode.Should().Be(HttpStatusCode.OK);
+        return;
 
         bool CheckRequest(HttpRequestMessage msg) {
             if (msg.Content is not MultipartFormDataContent formDataContent) return false;
 
-            return formDataContent.First().Headers.ContentType!.MediaType == "application/json";
+            return formDataContent.First().Headers.ContentType!.MediaType == ContentType.Json;
         }
     }
 }
