@@ -1,14 +1,16 @@
-namespace RestSharp.Tests; 
+namespace RestSharp.Tests;
 
 public class RequestHeaderTests {
+    static readonly KeyValuePair<string, string> AcceptHeader         = new(KnownHeaders.Accept, ContentType.Json);
+    static readonly KeyValuePair<string, string> AcceptLanguageHeader = new(KnownHeaders.AcceptLanguage, "en-us,en;q=0.5");
+    static readonly KeyValuePair<string, string> KeepAliveHeader      = new(KnownHeaders.KeepAlive, "300");
+
+    readonly List<KeyValuePair<string, string>> _headers = [AcceptHeader, AcceptLanguageHeader, KeepAliveHeader];
+
     [Fact]
     public void AddHeaders_SameCaseDuplicatesExist_ThrowsException() {
-        var headers = new List<KeyValuePair<string, string>> {
-            new("Accept", "application/json"),
-            new("Accept-Language", "en-us,en;q=0.5"),
-            new("Keep-Alive", "300"),
-            new("Accept", "application/json")
-        };
+        var headers = _headers;
+        _headers.Add(AcceptHeader);
 
         var request = new RestRequest();
 
@@ -18,12 +20,8 @@ public class RequestHeaderTests {
 
     [Fact]
     public void AddHeaders_DifferentCaseDuplicatesExist_ThrowsException() {
-        var headers = new List<KeyValuePair<string, string>> {
-            new("Accept", "application/json"),
-            new("Accept-Language", "en-us,en;q=0.5"),
-            new("Keep-Alive", "300"),
-            new("acCEpt", "application/json")
-        };
+        var headers = _headers;
+        headers.Add(new KeyValuePair<string, string>(KnownHeaders.Accept, ContentType.Json));
 
         var request = new RestRequest();
 
@@ -33,33 +31,25 @@ public class RequestHeaderTests {
 
     [Fact]
     public void AddHeaders_NoDuplicatesExist_Has3Headers() {
-        var headers = new List<KeyValuePair<string, string>> {
-            new("Accept", "application/json"),
-            new("Accept-Language", "en-us,en;q=0.5"),
-            new("Keep-Alive", "300")
-        };
-
         var request = new RestRequest();
-        request.AddHeaders(headers);
+        request.AddHeaders(_headers);
 
-        var httpParameters = request.Parameters.Where(parameter => parameter.Type == ParameterType.HttpHeader);
-
+        var httpParameters = GetHeaders(request);
         Assert.Equal(3, httpParameters.Count());
     }
 
     [Fact]
     public void AddHeaders_NoDuplicatesExistUsingDictionary_Has3Headers() {
         var headers = new Dictionary<string, string> {
-            { "Accept", "application/json" },
-            { "Accept-Language", "en-us,en;q=0.5" },
-            { "Keep-Alive", "300" }
+            { KnownHeaders.Accept, ContentType.Json },
+            { KnownHeaders.AcceptLanguage, "en-us,en;q=0.5" },
+            { KnownHeaders.KeepAlive, "300" }
         };
 
         var request = new RestRequest();
         request.AddHeaders(headers);
 
-        var httpParameters = request.Parameters.Where(parameter => parameter.Type == ParameterType.HttpHeader);
-
+        var httpParameters = GetHeaders(request);
         Assert.Equal(3, httpParameters.Count());
     }
 
@@ -67,16 +57,14 @@ public class RequestHeaderTests {
     public void AddOrUpdateHeader_ShouldUpdateExistingHeader_WhenHeaderExist() {
         // Arrange
         var request = new RestRequest();
-        request.AddHeader("Accept", "application/xml");
+        request.AddHeader(KnownHeaders.Accept, ContentType.Xml);
 
         // Act
-        request.AddOrUpdateHeader("Accept", "application/json");
+        request.AddOrUpdateHeader(KnownHeaders.Accept, ContentType.Json);
 
         // Assert
-        var headers = request.Parameters.Where(parameter => parameter.Type == ParameterType.HttpHeader).ToArray();
-
-        Assert.Equal("application/json", headers.First(parameter => parameter.Name == "Accept").Value);
-        Assert.Single(headers);
+        var headers = GetHeaders(request);
+        GetHeader(request, KnownHeaders.Accept).Should().Be(ContentType.Json);
     }
 
     [Fact]
@@ -85,22 +73,19 @@ public class RequestHeaderTests {
         var request = new RestRequest();
 
         // Act
-        request.AddOrUpdateHeader("Accept", "application/json");
+        request.AddOrUpdateHeader(KnownHeaders.Accept, ContentType.Json);
 
         // Assert
-        var headers = request.Parameters.Where(parameter => parameter.Type == ParameterType.HttpHeader).ToArray();
-
-        Assert.Equal("application/json", headers.First(parameter => parameter.Name == "Accept").Value);
-        Assert.Single(headers);
+        GetHeader(request, KnownHeaders.Accept).Should().Be(ContentType.Json);
     }
 
     [Fact]
     public void AddOrUpdateHeaders_ShouldAddHeaders_WhenNoneExists() {
         // Arrange
         var headers = new Dictionary<string, string> {
-            { "Accept", "application/json" },
-            { "Accept-Language", "en-us,en;q=0.5" },
-            { "Keep-Alive", "300" }
+            { KnownHeaders.Accept, ContentType.Json },
+            { KnownHeaders.AcceptLanguage, "en-us,en;q=0.5" },
+            { KnownHeaders.KeepAlive, "300" }
         };
 
         var request = new RestRequest();
@@ -109,25 +94,23 @@ public class RequestHeaderTests {
         request.AddOrUpdateHeaders(headers);
 
         // Assert
-        var requestHeaders = request.Parameters.Where(parameter => parameter.Type == ParameterType.HttpHeader).ToArray();
+        var requestHeaders = GetHeaders(request);
 
-        Assert.Equal("application/json", requestHeaders.First(parameter => parameter.Name == "Accept").Value);
-        Assert.Equal("en-us,en;q=0.5", requestHeaders.First(parameter => parameter.Name == "Accept-Language").Value);
-        Assert.Equal("300", requestHeaders.First(parameter => parameter.Name == "Keep-Alive").Value);
-        Assert.Equal(3, requestHeaders.Length);
+        var expected = headers.Select(x => new HeaderParameter(x.Key, x.Value));
+        expected.Should().BeEquivalentTo(requestHeaders);
     }
 
     [Fact]
     public void AddOrUpdateHeaders_ShouldUpdateHeaders_WhenAllExists() {
         // Arrange
         var headers = new Dictionary<string, string> {
-            { "Accept", "application/json" },
-            { "Keep-Alive", "300" }
+            { KnownHeaders.Accept, ContentType.Json },
+            { KnownHeaders.KeepAlive, "300" }
         };
 
         var updatedHeaders = new Dictionary<string, string> {
-            { "Accept", "application/xml" },
-            { "Keep-Alive", "400" }
+            { KnownHeaders.Accept, ContentType.Xml },
+            { KnownHeaders.KeepAlive, "400" }
         };
 
         var request = new RestRequest();
@@ -137,24 +120,25 @@ public class RequestHeaderTests {
         request.AddOrUpdateHeaders(updatedHeaders);
 
         // Assert
-        var requestHeaders = request.Parameters.Where(parameter => parameter.Type == ParameterType.HttpHeader).ToArray();
-
-        Assert.Equal("application/xml", requestHeaders.First(parameter => parameter.Name == "Accept").Value);
-        Assert.Equal("400", requestHeaders.First(parameter => parameter.Name == "Keep-Alive").Value);
-        Assert.Equal(2, requestHeaders.Length);
+        var requestHeaders = GetHeaders(request);
+        HeaderParameter[] expected = [
+            new HeaderParameter(KnownHeaders.Accept, ContentType.Xml),
+            new HeaderParameter(KnownHeaders.KeepAlive, "400")
+        ];
+        requestHeaders.Should().BeEquivalentTo(expected);
     }
 
     [Fact]
     public void AddOrUpdateHeaders_ShouldAddAndUpdateHeaders_WhenSomeExists() {
         // Arrange
         var headers = new Dictionary<string, string> {
-            { "Accept", "application/json" },
-            { "Keep-Alive", "300" }
+            { KnownHeaders.Accept, ContentType.Json },
+            { KnownHeaders.KeepAlive, "300" }
         };
 
         var updatedHeaders = new Dictionary<string, string> {
-            { "Accept", "application/xml" },
-            { "Accept-Language", "en-us,en;q=0.5" }
+            { KnownHeaders.Accept, ContentType.Xml },
+            { KnownHeaders.AcceptLanguage, "en-us,en;q=0.5" }
         };
 
         var request = new RestRequest();
@@ -164,11 +148,17 @@ public class RequestHeaderTests {
         request.AddOrUpdateHeaders(updatedHeaders);
 
         // Assert
-        var requestHeaders = request.Parameters.Where(parameter => parameter.Type == ParameterType.HttpHeader).ToArray();
+        var requestHeaders = GetHeaders(request);
 
-        Assert.Equal("application/xml", requestHeaders.First(parameter => parameter.Name == "Accept").Value);
-        Assert.Equal("en-us,en;q=0.5", requestHeaders.First(parameter => parameter.Name == "Accept-Language").Value);
-        Assert.Equal("300", requestHeaders.First(parameter => parameter.Name == "Keep-Alive").Value);
-        Assert.Equal(3, requestHeaders.Length);
+        HeaderParameter[] expected = [
+            new HeaderParameter(KnownHeaders.Accept, ContentType.Xml),
+            new HeaderParameter(KnownHeaders.AcceptLanguage, "en-us,en;q=0.5"),
+            new HeaderParameter(KnownHeaders.KeepAlive, "300")
+        ];
+        requestHeaders.Should().BeEquivalentTo(expected);
     }
+
+    static Parameter[] GetHeaders(RestRequest request) => request.Parameters.Where(x => x.Type == ParameterType.HttpHeader).ToArray();
+
+    static string GetHeader(RestRequest request, string name) => request.Parameters.FirstOrDefault(x => x.Name == name)?.Value?.ToString();
 }
